Refactor filereader function into promises
