+++ src/redux/actions/samples/createSample.js
@@ -47,6 +47,8 @@ const createSample = (
     },
   });
 
+  console.log('newSampleUuidDebug');
+  console.log(newSampleUuid);
   return Promise.resolve(newSampleUuid);
 };
